Corrects and refactors hard-coded "newbie" name check
WizardSupport.java prompts the user for a "real name" and password
if the user has not yet supplied them for the current game module.
The user is also prompted if his name is "newbie"

This check occurs inconsistently in the WizardSupport class.

- One of the checks the name against the i18n
  Resources.getString("Prefs.newbie").
- The other checks against a hard-coded "newbie" string.

This commit corrects and refactors the code to consistently
use the i18n string.