Fix JavaDoc search feature
This commit fixes "make javadoc" so that it generates pages that
return correct search results.

PROBLEM DESCRIPTION

"make javadoc" generates JavaDoc HTML documentation from the VASSAL
source code. This documentation includes a SEARCH field in the
upper-right corner of each page.  After filling in a search term in a
browser and clicking on a search result, the browser follows a
broken link. The page *does* exist, the link is just broken.

Example of the relative portion of an erroneous link:

   javadoc/undefined/VASSAL/command/AddPiece.html

Correct link:

   javadoc/VASSAL/command/AddPiece.html

STEPS TO REPRODUCE PROBLEM

1. From a command line, change to the root VASSAL project directory
   and run "make javadoc"
2. Use a browser to view "javadoc/index.html"
3. Enter "AddPiece" in the SEARCH field on the page
4. Click on the top result, "VASSAL.command.AddPiece"

The browser returns "ERR_FILE_NOT_FOUND" or something to that effect