Reset the decks sortListener on exit
This fixes deck sorting not working on the next and
each subsequent "File->New Game". Said another way, deck sorting
works the first time, but not after a 'close' and then 'new game',
unless the user completely exits Vassal and restarts it.

The sortListener must be null to assign a hotkey during initialization.