Debian package
This commit adds the `debian` directory to the repository, and sets up
the packaging workflow to also produce a Debian package.  The Debian
package is architecture independent, and should be installable on most
Debian derivatives.

The code is _not_ build by Maven, because many of the used Maven
packages and plugins are not available in Debian and its derivatives.
Debian packages _must_ build cleanly on a Debian system.  Thus, the code
and documentation is compiled and build directly in the `debian/rules`
file.

The Debian package integrates Vassal in to the desktop, and includes an
enhanced launcher script, as well as a `man(6)` page, and indexed
documentation for more system integrations.

When a new Vassal release is made, then the `debian/changelog` file
should be updated too.  This can be done with the utility `dch`
available on all Debian and derivative systems.  Perhaps we can find a
way to automatise this in the future.