Fix for invalid filenames when extracting archive
When extracting archive members, fix up the file name so that it is
valid on the target platform. Files are extracted from the archive
when the embedded HTML viewer needs a file - e.g., an image.

This fix calculates a SHA1 hash-code as a hexidecimal string, and uses
that as the extracted filename, thus using only the characters 0-9 and
a-f (A-F).   Since the SHA1 hash-code is predicatable, it will remain
the same on every invocation, and images are not needlessly extracted
from the archive.   It also avoids invalid characters or filenames on
platforms that may enforce such things in their file-systems.

Closes #14290 and similar issues.