Make charts zoomable
This PR will make charts zoomable, so as to allow users to zoom in or
out on an (image) chart display.  This is useful in case the chart has
small sized text or is very large, increasing the overall accessibility
of modules.

The zoom feature can be disabled if a module designer prefers the old
behaviour.   If zoom is enabled, then the chart display will listen for
Ctrl-MouseWheel and configurable key events (`+` and `-`) by default.

The added configurables are
- whether to enable zoom all together (enabled by default)
- the least zoom factor (30% by default)
- the largest zoom factor (300% by default)
- the zoom in hotkey (`+` by default)
- the zoom out hotkey (`-` by default)

One could consider to add a tool-bar and zoom in and out buttons.

This partially closes #13035